refactor: clean up icon components for consistency
