+++ app/web_ui/src/lib/ui/icons/checkmark_icon.svelte
