+++ app/web_ui/src/lib/ui/icons/finetune_icon.svelte
@@ -1,6 +1,3 @@
-<script lang="ts">
-</script>
-
 <!-- Uploaded to: SVG Repo, www.svgrepo.com, Generator: SVG Repo Mixer Tools -->
 <svg
   class="w-full h-full"
