+++ app/web_ui/src/lib/ui/icons/qna_icon.svelte
@@ -1,5 +1,3 @@
-<script lang="ts"></script>
-
 <!-- Uploaded to: SVG Repo, www.svgrepo.com, Generator: SVG Repo Mixer Tools -->
 <svg
   class="w-full h-full"
