+++ app/web_ui/src/lib/ui/icons/trash_icon.svelte
@@ -1,6 +1,3 @@
-<script lang="ts">
-</script>
-
 <!-- Uploaded to: SVG Repo, www.svgrepo.com, Generator: SVG Repo Mixer Tools -->
 <svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
   <path
